requirements.txt
pandas
numpy
pdfplumber
jupyter
PyMuPDF
Pillow
langchain
sentence-transformers
qdrant-client
openai
tiktoken
langchainhub
python-dotenv